move query modules
