+++ crates/lib/mysql_queries/src/queries/voice_designer/mod.rs
@@ -1,5 +1,3 @@
-pub mod create_dataset;
-pub mod update_dataset;
-pub mod get_dataset;
-pub mod delete_dataset;
-pub mod create_voice;
+pub mod datasets;
+pub mod voice_samples;
+pub mod voices;
