+++ crates/lib/mysql_queries/src/queries/voice_designer/voices/create_voice.rs
