+++ crates/lib/mysql_queries/src/queries/voice_designer/voices/mod.rs
@@ -1,0 +1,1 @@
+pub mod create_voice;
