+++ crates/service/web/storyteller_web/src/http_server/endpoints/voice_designer/delete_dataset.rs
@@ -10,8 +10,8 @@ use enums::by_table::audit_logs::audit_log_entity_action::AuditLogEntityAction::
 
 use http_server_common::request::get_request_ip::get_request_ip;
 use http_server_common::response::serialize_as_json_error::serialize_as_json_error;
-use mysql_queries::queries::voice_designer::delete_dataset::{delete_dataset_as_mod, delete_dataset_as_user, undelete_dataset_as_mod, undelete_dataset_as_user};
-use mysql_queries::queries::voice_designer::get_dataset::get_dataset_by_token;
+use mysql_queries::queries::voice_designer::datasets::delete_dataset::{delete_dataset_as_mod, delete_dataset_as_user, undelete_dataset_as_mod, undelete_dataset_as_user};
+use mysql_queries::queries::voice_designer::datasets::get_dataset::get_dataset_by_token;
 use tokens::tokens::zs_dataset::ZsDatasetToken;
 use crate::http_server::web_utils::response_success_helpers::simple_json_success;
